Moved sunspot indexer into appropriate namespace
